Reorganize files in panther_battery
