+++ panther_battery/CMakeLists.txt
@@ -18,11 +18,11 @@ add_executable(
   battery_node
   src/main.cpp
   src/battery_node.cpp
-  src/adc_battery.cpp
-  src/roboteq_battery.cpp
-  src/battery_publisher.cpp
-  src/dual_battery_publisher.cpp
-  src/single_battery_publisher.cpp)
+  src/battery/adc_battery.cpp
+  src/battery/roboteq_battery.cpp
+  src/battery_publisher/battery_publisher.cpp
+  src/battery_publisher/dual_battery_publisher.cpp
+  src/battery_publisher/single_battery_publisher.cpp)
 ament_target_dependencies(battery_node ${PACKAGE_DEPENDENCIES})
 
 install(TARGETS battery_node DESTINATION lib/${PROJECT_NAME})
@@ -39,7 +39,7 @@ if(BUILD_TESTING)
     PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
            $<INSTALL_INTERFACE:include>)
 
-  ament_add_gtest(${PROJECT_NAME}_test_battery test/test_battery.cpp)
+  ament_add_gtest(${PROJECT_NAME}_test_battery test/battery/test_battery.cpp)
   target_include_directories(
     ${PROJECT_NAME}_test_battery
     PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
@@ -47,8 +47,8 @@ if(BUILD_TESTING)
   ament_target_dependencies(${PROJECT_NAME}_test_battery rclcpp sensor_msgs
                             panther_msgs)
 
-  ament_add_gtest(${PROJECT_NAME}_test_adc_battery test/test_adc_battery.cpp
-                  src/adc_battery.cpp)
+  ament_add_gtest(${PROJECT_NAME}_test_adc_battery
+                  test/battery/test_adc_battery.cpp src/battery/adc_battery.cpp)
   target_include_directories(
     ${PROJECT_NAME}_test_adc_battery
     PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
@@ -56,8 +56,9 @@ if(BUILD_TESTING)
   ament_target_dependencies(${PROJECT_NAME}_test_adc_battery rclcpp sensor_msgs
                             panther_msgs)
 
-  ament_add_gtest(${PROJECT_NAME}_test_roboteq_battery
-                  test/test_roboteq_battery.cpp src/roboteq_battery.cpp)
+  ament_add_gtest(
+    ${PROJECT_NAME}_test_roboteq_battery test/battery/test_roboteq_battery.cpp
+    src/battery/roboteq_battery.cpp)
   target_include_directories(
     ${PROJECT_NAME}_test_roboteq_battery
     PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
@@ -65,8 +66,10 @@ if(BUILD_TESTING)
   ament_target_dependencies(${PROJECT_NAME}_test_roboteq_battery panther_msgs
                             rclcpp sensor_msgs)
 
-  ament_add_gtest(${PROJECT_NAME}_test_battery_publisher
-                  test/test_battery_publisher.cpp src/battery_publisher.cpp)
+  ament_add_gtest(
+    ${PROJECT_NAME}_test_battery_publisher
+    test/battery_publisher/test_battery_publisher.cpp
+    src/battery_publisher/battery_publisher.cpp)
   target_include_directories(
     ${PROJECT_NAME}_test_battery_publisher
     PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
@@ -76,8 +79,10 @@ if(BUILD_TESTING)
 
   ament_add_gtest(
     ${PROJECT_NAME}_test_single_battery_publisher
-    test/test_single_battery_publisher.cpp src/adc_battery.cpp
-    src/battery_publisher.cpp src/single_battery_publisher.cpp)
+    test/battery_publisher/test_single_battery_publisher.cpp
+    src/battery/adc_battery.cpp
+    src/battery_publisher/battery_publisher.cpp
+    src/battery_publisher/single_battery_publisher.cpp)
   target_include_directories(
     ${PROJECT_NAME}_test_single_battery_publisher
     PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
@@ -87,8 +92,10 @@ if(BUILD_TESTING)
 
   ament_add_gtest(
     ${PROJECT_NAME}_test_dual_battery_publisher
-    test/test_dual_battery_publisher.cpp src/adc_battery.cpp
-    src/battery_publisher.cpp src/dual_battery_publisher.cpp)
+    test/battery_publisher/test_dual_battery_publisher.cpp
+    src/battery/adc_battery.cpp
+    src/battery_publisher/battery_publisher.cpp
+    src/battery_publisher/dual_battery_publisher.cpp)
   target_include_directories(
     ${PROJECT_NAME}_test_dual_battery_publisher
     PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
@@ -97,51 +104,51 @@ if(BUILD_TESTING)
                             ${PACKAGE_DEPENDENCIES})
 
   ament_add_gtest(
-    ${PROJECT_NAME}_test_battery_node
-    test/test_battery_node.cpp
-    src/adc_battery.cpp
+    ${PROJECT_NAME}_test_battery_node_adc_dual
+    test/test_battery_node_adc_dual.cpp
+    src/battery/adc_battery.cpp
     src/battery_node.cpp
-    src/battery_publisher.cpp
-    src/dual_battery_publisher.cpp
-    src/single_battery_publisher.cpp
-    src/roboteq_battery.cpp)
+    src/battery_publisher/battery_publisher.cpp
+    src/battery_publisher/dual_battery_publisher.cpp
+    src/battery_publisher/single_battery_publisher.cpp
+    src/battery/roboteq_battery.cpp)
   target_include_directories(
-    ${PROJECT_NAME}_test_battery_node
+    ${PROJECT_NAME}_test_battery_node_adc_dual
     PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/test>
            $<INSTALL_INTERFACE:include>)
-  ament_target_dependencies(${PROJECT_NAME}_test_battery_node
+  ament_target_dependencies(${PROJECT_NAME}_test_battery_node_adc_dual
+                            ${PACKAGE_DEPENDENCIES})
+
+  ament_add_gtest(
+    ${PROJECT_NAME}_test_battery_node_adc_single
+    test/test_battery_node_adc_single.cpp
+    src/battery/adc_battery.cpp
+    src/battery_node.cpp
+    src/battery_publisher/battery_publisher.cpp
+    src/battery_publisher/dual_battery_publisher.cpp
+    src/battery_publisher/single_battery_publisher.cpp
+    src/battery/roboteq_battery.cpp)
+  target_include_directories(
+    ${PROJECT_NAME}_test_battery_node_adc_single
+    PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/test>
+           $<INSTALL_INTERFACE:include>)
+  ament_target_dependencies(${PROJECT_NAME}_test_battery_node_adc_single
                             ${PACKAGE_DEPENDENCIES})
 
   ament_add_gtest(
     ${PROJECT_NAME}_test_battery_node_roboteq
     test/test_battery_node_roboteq.cpp
-    src/adc_battery.cpp
+    src/battery/adc_battery.cpp
     src/battery_node.cpp
-    src/battery_publisher.cpp
-    src/dual_battery_publisher.cpp
-    src/single_battery_publisher.cpp
-    src/roboteq_battery.cpp)
+    src/battery_publisher/battery_publisher.cpp
+    src/battery_publisher/dual_battery_publisher.cpp
+    src/battery_publisher/single_battery_publisher.cpp
+    src/battery/roboteq_battery.cpp)
   target_include_directories(
     ${PROJECT_NAME}_test_battery_node_roboteq
     PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/test>
            $<INSTALL_INTERFACE:include>)
   ament_target_dependencies(${PROJECT_NAME}_test_battery_node_roboteq
-                            ${PACKAGE_DEPENDENCIES})
-
-  ament_add_gtest(
-    ${PROJECT_NAME}_test_battery_node_dual_bat
-    test/test_battery_node_dual_bat.cpp
-    src/adc_battery.cpp
-    src/battery_node.cpp
-    src/battery_publisher.cpp
-    src/dual_battery_publisher.cpp
-    src/single_battery_publisher.cpp
-    src/roboteq_battery.cpp)
-  target_include_directories(
-    ${PROJECT_NAME}_test_battery_node_dual_bat
-    PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/test>
-           $<INSTALL_INTERFACE:include>)
-  ament_target_dependencies(${PROJECT_NAME}_test_battery_node_dual_bat
                             ${PACKAGE_DEPENDENCIES})
 
 endif()
