+++ panther_battery/src/battery/roboteq_battery.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "panther_battery/roboteq_battery.hpp"
+#include "panther_battery/battery/roboteq_battery.hpp"
 
 #include <cstdint>
 #include <functional>
