+++ panther_battery/test/battery_publisher/test_single_battery_publisher.cpp
@@ -21,9 +21,9 @@
 
 #include "sensor_msgs/msg/battery_state.hpp"
 
-#include "panther_battery/adc_battery.hpp"
-#include "panther_battery/battery.hpp"
-#include "panther_battery/single_battery_publisher.hpp"
+#include "panther_battery/battery/adc_battery.hpp"
+#include "panther_battery/battery/battery.hpp"
+#include "panther_battery/battery_publisher/single_battery_publisher.hpp"
 #include "panther_utils/test/ros_test_utils.hpp"
 
 using BatteryStateMsg = sensor_msgs::msg::BatteryState;
