+++ panther_battery/test/test_battery_node_adc_single.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "include/test_battery_node.hpp"
+#include "utils/test_battery_node.hpp"
 
 #include <chrono>
 
@@ -22,67 +22,39 @@
 
 #include "panther_utils/test/ros_test_utils.hpp"
 
-class TestBatteryNodeDualBattery : public TestBatteryNode
+class TestBatteryNodeADCSingle : public TestBatteryNode
 {
 public:
-  TestBatteryNodeDualBattery() : TestBatteryNode(1.2, true) {}
+  TestBatteryNodeADCSingle() : TestBatteryNode(1.2, false) {}
 };
 
-TEST_F(TestBatteryNodeDualBattery, BatteryValues)
+TEST_F(TestBatteryNodeADCSingle, BatteryValues)
 {
+  // Change some values for ADC channels used by second battery.
+  // As this is test for single battery this should not affect published battery state.
+  WriteNumberToFile<int>(1600, std::filesystem::path(device1_path_ / "in_voltage3_raw"));
+  WriteNumberToFile<int>(100, std::filesystem::path(device0_path_ / "in_voltage2_raw"));
+
   ASSERT_TRUE(panther_utils::test_utils::WaitForMsg(
     battery_node_, battery_state_, std::chrono::milliseconds(5000)));
 
   // This is done to check if channels of ADC readers were assigned correctly, not to verify
   // calculations. If any test performing calculations fails this test will most likely fail too.
   EXPECT_FLOAT_EQ(35.05957, battery_state_->voltage);
-  EXPECT_FLOAT_EQ(2.02, battery_state_->current);
+  EXPECT_FLOAT_EQ(2.01, battery_state_->current);
   EXPECT_FLOAT_EQ(26.094543, battery_state_->temperature);
   EXPECT_FLOAT_EQ(0.32548615, battery_state_->percentage);
-  EXPECT_FLOAT_EQ(13.019446, battery_state_->charge);
+  EXPECT_FLOAT_EQ(6.5097232, battery_state_->charge);
 
-  // If both batteries have the same reading values they should have equal values
-  EXPECT_FLOAT_EQ(battery_1_state_->voltage, battery_2_state_->voltage);
-  EXPECT_FLOAT_EQ(battery_1_state_->current, battery_2_state_->current);
-  EXPECT_FLOAT_EQ(battery_1_state_->temperature, battery_2_state_->temperature);
-  EXPECT_FLOAT_EQ(battery_1_state_->percentage, battery_2_state_->percentage);
-  EXPECT_FLOAT_EQ(battery_1_state_->charge, battery_2_state_->charge);
-
-  // Change value of battery 2 reading one by one and check if corresponding values in battery 1
-  // stops matching
-  WriteNumberToFile<int>(1600, std::filesystem::path(device1_path_ / "in_voltage3_raw"));
-  ASSERT_TRUE(panther_utils::test_utils::WaitForMsg(
-    battery_node_, battery_2_state_, std::chrono::milliseconds(1000)));
-  EXPECT_FALSE(
-    fabs(battery_1_state_->voltage - battery_2_state_->voltage) <
-    std::numeric_limits<float>::epsilon());
-  EXPECT_FALSE(
-    fabs(battery_1_state_->percentage - battery_2_state_->percentage) <
-    std::numeric_limits<float>::epsilon());
-  EXPECT_FALSE(
-    fabs(battery_1_state_->charge - battery_2_state_->charge) <
-    std::numeric_limits<float>::epsilon());
-  EXPECT_FLOAT_EQ(battery_1_state_->current, battery_2_state_->current);
-  EXPECT_FLOAT_EQ(battery_1_state_->temperature, battery_2_state_->temperature);
-
-  WriteNumberToFile<int>(900, std::filesystem::path(device1_path_ / "in_voltage1_raw"));
-  WriteNumberToFile<int>(100, std::filesystem::path(device0_path_ / "in_voltage2_raw"));
-  ASSERT_TRUE(panther_utils::test_utils::WaitForMsg(
-    battery_node_, battery_2_state_, std::chrono::milliseconds(1000)));
-  EXPECT_FALSE(
-    fabs(battery_1_state_->current - battery_2_state_->current) <
-    std::numeric_limits<float>::epsilon());
-  EXPECT_FLOAT_EQ(battery_1_state_->temperature, battery_2_state_->temperature);
-
-  WriteNumberToFile<int>(1000, std::filesystem::path(device0_path_ / "in_voltage0_raw"));
-  ASSERT_TRUE(panther_utils::test_utils::WaitForMsg(
-    battery_node_, battery_2_state_, std::chrono::milliseconds(1000)));
-  EXPECT_FALSE(
-    fabs(battery_1_state_->temperature - battery_2_state_->temperature) <
-    std::numeric_limits<float>::epsilon());
+  // For single battery if readings stay the same values of battery_1 and battery should be the same
+  EXPECT_FLOAT_EQ(battery_1_state_->voltage, battery_state_->voltage);
+  EXPECT_FLOAT_EQ(battery_1_state_->current, battery_state_->current);
+  EXPECT_FLOAT_EQ(battery_1_state_->temperature, battery_state_->temperature);
+  EXPECT_FLOAT_EQ(battery_1_state_->percentage, battery_state_->percentage);
+  EXPECT_FLOAT_EQ(battery_1_state_->charge, battery_state_->charge);
 }
 
-TEST_F(TestBatteryNodeDualBattery, BatteryTimeout)
+TEST_F(TestBatteryNodeADCSingle, BatteryTimeout)
 {
   ASSERT_TRUE(panther_utils::test_utils::WaitForMsg(
     battery_node_, battery_state_, std::chrono::milliseconds(5000)));
@@ -111,7 +83,7 @@ TEST_F(TestBatteryNodeDualBattery, BatteryTimeout)
   EXPECT_EQ(BatteryStateMsg::POWER_SUPPLY_HEALTH_UNKNOWN, battery_state_->power_supply_health);
 }
 
-TEST_F(TestBatteryNodeDualBattery, BatteryCharging)
+TEST_F(TestBatteryNodeADCSingle, BatteryCharging)
 {
   // Wait for node to initialize
   ASSERT_TRUE(panther_utils::test_utils::WaitForMsg(
@@ -125,6 +97,30 @@ TEST_F(TestBatteryNodeDualBattery, BatteryCharging)
     battery_node_, battery_state_, std::chrono::milliseconds(1000)));
 
   EXPECT_NE(battery_state_->power_supply_status, BatteryStateMsg::POWER_SUPPLY_STATUS_DISCHARGING);
+}
+
+TEST_F(TestBatteryNodeADCSingle, RoboteqInitOnADCFail)
+{
+  // Remove ADC device
+  std::filesystem::remove_all(device0_path_);
+
+  // Wait for node to initialize
+  ASSERT_TRUE(panther_utils::test_utils::WaitForMsg(
+    battery_node_, battery_state_, std::chrono::milliseconds(5000)));
+
+  // Battery state status should be UNKNOWN
+  EXPECT_EQ(BatteryStateMsg::POWER_SUPPLY_STATUS_UNKNOWN, battery_state_->power_supply_status);
+
+  // Publish driver state that should update the battery message
+  DriverStateMsg driver_state;
+  driver_state.header.stamp = battery_node_->get_clock()->now();
+  driver_state_pub_->publish(driver_state);
+
+  ASSERT_TRUE(panther_utils::test_utils::WaitForMsg(
+    battery_node_, battery_state_, std::chrono::milliseconds(1000)));
+
+  // Battery state status should be different than UNKNOWN
+  EXPECT_NE(BatteryStateMsg::POWER_SUPPLY_STATUS_UNKNOWN, battery_state_->power_supply_status);
 }
 
 int main(int argc, char ** argv)
