+++ panther_battery/test/test_battery_node_roboteq.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "include/test_battery_node.hpp"
+#include "utils/test_battery_node.hpp"
 
 #include <chrono>
 #include <cmath>
